fix to simple_wml on 64 bit machines

+++ src/server/simple_wml.hpp
@@ -146,7 +146,7 @@ private:
 
 	void set_dirty();
 
-	void shift_buffers(int offset);
+	void shift_buffers(ptrdiff_t offset);
 
 	document* doc_;
